Added Validations to the Remove and Edit comments method to make sure the comment belongs to the user

+++ src/main/java/com/app/service/CommentService.java
@@ -12,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 
 
 @Service
@@ -21,6 +22,7 @@ public class CommentService implements CommentUseCases{
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
     private final PostRepository postRepository;
+    private final UserUseCases userUseCases;
 
     public void validateComment(String text) {
         if (text == null || text.isBlank()) {
@@ -65,13 +67,13 @@ public class CommentService implements CommentUseCases{
     @Transactional
     public void editComment(long commentId, String newText) {
         Comment comment = findCommentById(commentId);
-        if (comment != null) {
-            validateComment(newText);
-            comment.setText(newText);
-            commentRepository.save(comment);
-        } else {
-            throw new IllegalArgumentException("Comment with id " + commentId + " not found");
+
+        if (!Objects.equals(comment.getUser().getId(), userUseCases.getLoggedInUser().getId())) {
+            throw new IllegalStateException("This comment was not created by You ");
         }
+        validateComment(newText);
+        comment.setText(newText);
+        commentRepository.save(comment);
     }
 
     @Transactional
@@ -79,8 +81,9 @@ public class CommentService implements CommentUseCases{
         // try to find this comment
         Comment comment = findCommentById(commentId);
 
-        // if not found, method throws exception
-        // if found, we delete
+        if (!Objects.equals(comment.getUser().getId(), userUseCases.getLoggedInUser().getId())) {
+            throw new IllegalStateException("This comment was not created by You ");
+        }
         commentRepository.deleteById(commentId);
     }
 }
